[runtime] Integrate snitch runtime into the quidditch runtime
Building the snitch runtime as part of the monobuild ensures that we always use the same compiler, compilation flags and also gives us greater flexibility in assembling the parts from the snitch runtime that we need. In particular, changing the startup file to differently use the L1 of the compute cores will likely be required.
Lastly, this makes it very easy to change the hardware config as it only requires changing the path to a hjson file within the cmake.

It is now possible to build and run a hello world as a check.