[runtime] Add basic xDSL kernel compilation example (#5)
This PR implements a basic sample running an xDSL generated kernel in a
snitch cluster. The code currently has to be executed by a compute core
(as the kernels are executed by the same caller core) and the memory
allocated in L1 as that is required by streaming registers. The
quidditch device was reverted to just be a fork of the `local-sync`
device as the current plan is to leverage parallelism inside one cluster
within a kernel for now.