[dispatch] Synchronize using barriers rather than local interrupts
The `wfi` instruction, technically speaking, is not guaranteed to halt the core nor be woken up by exactly the interrupt we want. Hardware barriers are the more canonical way of implementing synchronization within a cluster and our new implementation also parks the DMA core during kernel execution instead of making it busy loop. This is mainly useful for simulation time and power usage in real code.