[compiler][runtime] Compile modules with both LLVM and Quidditch
Having both Quidditch and LLVM output allows us to freely use kernels generated by either of the two, depending on the order they're registered in our loader.

The main use-case is for us to be able to gradually introduce xDSL output by falling back to LLVM when xDSL is incapable of compiling a kernel.