[compiler][NFC] Move `xDSL` conversion to RISC-V into a pass
This moves the conversion into the translation pipeline rather than during serialization.
Our use-case for this is that we want to perform actions in the linker pass pipeline run afterward based on the output of the xDSL conversion.
It additionally makes the code more readable by moving logic out of the not small QuidditchTarget.cpp file.