[compiler] use an exclusive by default outlining pass
The previous pass that performed outlining functions for xDSL was very ambitious by outlining everything that wasn't IREE's HAL dialect to `xDSL`. This inclusive by default approach was rather ambitious as `xDSL` is not capable of lowering a lot of things to RISC-V.

This PR therefore reimplements that pass but as an exclusive by default approach. For the time being, only a series of linalg-on-memref operations are outlined into a function + their constants. This happens to also create single basic block functions.