[compiler] Add pattern to raise ReLUs in select to `arith.maximumf` form
The latter xDSL is capable of lowering to floating point instructions that can be placed inside a `frep` and is therefore a lot faster.

As a side-effect two more kernels in NsNet2 can be compiled which showed an issue with `hal.executable.condition` in the linker that had to be fixed as well.