[compiler] Introduce `quidditch_snitch` dialect
We are starting to introduce more snitch and xDSL specific concepts into the codegen pipeline that we'll want to encapsulate into operations that are progressively lowered and optimized.
As a simple start, the dialect adds the `xdsl_kernel` op which contains linalg operations that are to be compiled by xDSL.
Many discardable attributes that we've been using can now also be represented as dialect attribute.