[quidditch_snitch] Implement bufferization for microkernels (#38)
By making it possible to bufferize the operations we are able to form
our micro kernels earlier in the pass pipeline prior to bufferization.
This is required for other transformation passes such as explicit tensor
promotion to L1 memory and subgroup distribution that we are planning to
add later.

As a consequence of this change the existing `xdsl_kernel` (now renamed
to just `microkernel`) had to be split into two operations, one
operating on tensors, the other on memrefs. The reason for this is that
the memref version should be isolated from above while the tensor
version cannot be for bufferization purposes. The bufferization process
itself now converts between the two.