[quidditch_snitch] Implement tensor promotion to L1
An invariant of our microkernels is that tensors MUST be in L1 memory or the execution will crash.
This PR implements enforces this by adding and creating a new operation called `copy_l1_tensor` which is capable of transferring either from or to L1.
During bufferization a copy is performed, if necessary, otherwise the operation is a noop.
The `L1EncodingAttr` is used as memory space for L1 MemRefs during bufferization whose lowerings have been implemented in previous PRs.

With this PR it is now also possible to run kernels with buffers allocated in L3/DRAM as long as the used memory fits into L1.