[compiler] Satisfy invariant of `hal.executable.export`
`hal.executable.export` has the (reasonable) invariant of the symbol that is being exported referring to the function that serves as the kernel entry point. We previously broke this invariant by renaming the entrypoint.
This PR moves the renaming to very late in the pipeline in LLVM IR to uphold the invariant while in MLIR. A side-effect of this change is that we completely stop depending on the name of the function having any significance on semantics.