[compiler] Implement Reduction and Thread tiling
Both levels of tiling are cruical optimizations for accelerating the NN execution on snitch. Reduction tiling allows us to make larger workgroups and increase parallism for the later thread tiling.
Thread tiling is required to distribute a linalg operations accross all 8 compute cores.

The former materialzies as a `scf.for` while the latter to an `scf.forall`. To make this PR smaller, the `scf.forall` is lowered to just an `scf.for` with later PRs implementing the distribution of `scf.forall`s.