[compiler][runtime] Implement distributing `forall` to all compute cores
A previous PR implemented tiling of parallel dimensions to `scf.forall` but did not yet implement the lowering of `scf.forall`.
This PR implements the lowering of `scf.forall` and distributing it to every compute core. This is achieved using the `quidditch_snitch` op which returns an id of 0 to `compute_cores - 1`. We then lower the `scf.forall` to an `scf.for` where each compute core takes an `1/compute_cores` slice of the iteration space.