[cmake] Mark python requirements 'CONFIGURE_DEPENDS' (#80)
Marking them as such causes cmake to reconfigure and rerun `pip install`
if any of these files changed. Most importantly, it also triggers if the
`xDSL` version we use changes, making it impossible to accidentally test
and use an xDSL installation different from the one checked out as a
submodule.