[TensorTile] Add dedicated L1 tiling
Previously, workgroup tiling was responsible for two things:
* Making sure the working set fits into L1 memory and
* Splitting into workgroups for distribution

However, this is undesirable as this forces us to choose a tile size satisfying both requirements.
Instead, this PR adds a new level of tiling called L1 tiling which takes over the responsibility of making the working set fit into L1 memory.
Workgroup distribution tiling is now solely responsible for workgroup concurrency.
As we are currently compiling for just one snitch cluster, the existing tile sizes have been changed to L1 tile sizes except the reduction dimension.
The reduction dimension tiling should come prior to L1 tiling, which happens to be the case right now.
Larger workgroups are desirable in the absence of concurrency as we avoid redundant data transfer of workgroup-id invariant data (input and ouput vectors in the mat-vec) + any launch overhead.