[Target] Fork LLVMCPU's `ConvertToLLVM` pass
Any patterns involving type system changes need to be run in one large dialect conversion. Our existing `ConvertSnitchToLLVM` pass has so far worked as it did not interact with other dialects operations but is nevertheless strongly limited.

This PR therefore does the initial step of forking the `ConvertToLLVM` pass from IREE's LLVMCPU backend. Only the bare minimum of changes for integration are performed to have a clean diff of Quidditch specific changes in follow up PRs.

This additionally creates exciting opportunities in the future regarding a more optimized ABI for Snitch and Occamy.