[quidditch_snitch] Implement `pipeline-copy-compute` pass
This PR implements the lifting of `scf.for` ops containing a compute op + tensor copy ops into a `pipeline` op where copy and compute operations are split into two pipeline stages.
An additional `dual_buffer` attribute on the lowering config of the compute op controls whether this should be performed (which it almost should be).
This also updates our configs and pipeline to perform dual buffering of the rows of the matrix.