[ConvertToRISCV] Fallback to Linalg lowering if `xdsl-opt` fails
While we are not able to leverage snitch's SSR and FREP extension if `xdsl-opt` fails, our schedule and use of L1 memory should generally lead to performance improvements over the LLVM backend's kernel.
This PR therefore changes the mechanism when `xdsl-opt` fails to inline the linalg operation and erase the `quidditch_snitch.memref.microkernel`, instead of the current behavior of discarding the kernel entirely.