[quidditch_snitch] Add `start_zero_mem_transfer` operation (#113)
This operation is a special kind of DMA transfer leveraging both
Snitch's DMA and a special address space in the cluster's memory that
always returns 0. This allows zeroing memory in bursts of 64kB using the
DMA.

While currently unused in the pipeline, the operation will be used soon
to implement a lowering of `tensor.pad`. Occurrences of `linalg.fill`
with zero can also be optimized in the future.