[PromotePadsToL1] Implement pass lowering `tensor.pad`
`tensor.pad`, as we produce them, can be lowered to `start_tensor_copy`. The latter will perform the L1 allocation and copy operations during bufferization.

The one not-nice part of the pass is that it does so indiscriminately for all `tensor.pad` operations. I am very unsure about whether there are any guarantees by IREE that'd never encounter `tensor.pad` operations in kernels.

This was also the last pass required to compile the last kernel in NsNet2