[SnitchDMA] Introduce `SnitchDMA` dialect
The DMA dialect represents DMA operations in their purest forms: Moving data from one MemRef to another.
The real world is unfortunately not as pretty: Snitch's DMA only supports two-dimensional transfers, the zero region is of a specific size and DMA configuration is relatively expensive.

This PR introduces the `SnitchDMA` dialect which is meant to handle any Snitch specific legalizations but also optimizations.

The intention for the future is to have lowering of DMA operations split into multiple passes and phases performing 1) legalization of DMA transfers, 2) Optimization of configuration, 3) optimization of number of waits and 4) lowering to LLVM.

As a first step, this PR only introduces the dialect and the equivalent of the `dmstati` instruction which is used in the current one-shot to LLVM lowering.