[DMA] Add `combine_token` op (#130)
Being able to dynamically combine tokens to await a runtime-dependent
number of DMA transfers is a requirement for implementing legalization
of DMA transfers in the `dma` dialect.

This PR therefore adds the `combined_tokens` op which combines multiple
tokens into one. The lowering for Snitch leverages the monotonicity
guarantee of IDs to combine them. Note that this only works with a
single channel DMA right now.

As this subsumes the multi-token capabilities of `wait_for_transfers`,
it has been renamed to just `wait_for_transfer` and only accepts a
single token input now.