[build] Disable use of logf128 to fix build failures (#138)
This PR:

- Disables the use (and auto-detection during build configuration) of
the natural logarithm on FP128 that caused a build failure within IREE's
bundled LLVM (this is related to this
[fix](https://github.com/llvm/llvm-project/issues/100296)).

### Description

IREE ships with a bundled LLVM within its `third_party` subdirectory to
allow code generation of various targets (e.g., RISC-V via the LLVM-CPU
backend). This bundled LLVM has limited configurability (mostly
hardcoded via
[`iree_llvm.cmake`](https://github.com/iree-org/iree/blob/456d80c51930ccc03ce0488e98238e5e0a14b403/build_tools/cmake/iree_llvm.cmake#L17)),
hence we rely on `cmake` to propagate this compile definition from the
top-level directory when building Quidditch's `codegen` sub-project.

This behaviour was first exhibited when the host compiler was updated
(on a local build host) to version 19.1.7.